Expose shell interface.
