+++ src/lib.rs
@@ -27,6 +27,7 @@ pub mod data_device;
 pub mod keyboard;
 pub mod output;
 pub mod pointer;
+pub mod shell;
 pub mod surface;
 pub mod utils;
 pub mod window;
